prj-629/pyproject.toml
[project]
name = "prj-629"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-cloud-bigquery>=3.39.0",
    "python-dotenv>=1.2.1",
]
